Fix for neutron test with id 843870
For VLAN configuration 29 instances can be created according to the
implementation. In test_check_port_binding_after_restart_node we try to
create 29 new instances, but some of them have been already created in
_prepare_openstack.
In this fix the total number of instance to be created is changed to 29.
stable/8.0 branch.

Change-Id: I5a9c486341b36c26a56798ceb61be219cbc70d0b